# api-docs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/aulas": {
      "post": {
        "operationId": "AulasController_criar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriarAulaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "A aula foi cadastrada com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AulaRespostaDto"
                }
              }
            }
          },
          "400": {
            "description": "A turma não está EM_ANDAMENTO ou a data enviada está fora do período letivo estabelecido."
          },
          "409": {
            "description": "Já existe uma aula cadastrada nesta mesma data para a turma informada."
          },
          "500": {
            "description": "Ocorreu um erro inesperado ao cadastrar a aula."
          }
        },
        "summary": "Cadastrar ou agendar uma nova aula para uma turma",
        "tags": [
          "Aulas"
        ]
      },
      "get": {
        "operationId": "AulasController_buscarTodas",
        "parameters": [
          {
            "name": "turmaId",
            "required": true,
            "in": "query",
            "description": "ID da turma para listar as aulas",
            "schema": {
              "example": "1",
              "type": "string"
            }
          },
          {
            "name": "pagina",
            "required": false,
            "in": "query",
            "description": "Número da página atual (padrão: 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "itensPorPagina",
            "required": false,
            "in": "query",
            "description": "Número de itens por página (padrão: 10)",
            "schema": {
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resposta paginada recebida com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginacaoRespostaDto"
                    },
                    {
                      "properties": {
                        "dados": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/AulaRespostaDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Os parâmetros de paginação (página ou itensPorPagina) devem ser maiores ou iguais a 1."
          },
          "500": {
            "description": "Ocorreu um erro inesperado ao buscar a listagem de aulas."
          }
        },
        "summary": "Buscar uma lista paginada de aulas registradas",
        "tags": [
          "Aulas"
        ]
      }
    },
    "/aulas/{id}": {
      "get": {
        "operationId": "AulasController_buscarPorId",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A aula foi encontrada com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AulaRespostaDto"
                }
              }
            }
          },
          "404": {
            "description": "Aula com o ID especificado não foi encontrada."
          },
          "500": {
            "description": "Ocorreu um erro inesperado ao buscar a aula."
          }
        },
        "summary": "Buscar o registro de uma aula por ID",
        "tags": [
          "Aulas"
        ]
      },
      "patch": {
        "description": "Permite atualizar o tema, a data e o status de uma aula existente.\n\n**Regras de Negócio e Transições de Status:**\n- **Presenças Registradas**: Uma aula com chamadas salvas não pode retornar ao status `AGENDADA` nem ser alterada para `CANCELADA` (é necessário excluir o lote de chamadas primeiro caso tenha sido feito por engano).\n- **Marcação de Realização**: Uma aula só pode ser marcada como `REALIZADA` se possuir chamadas registradas.\n- **Limites do Calendário**: A data da aula deve estar estritamente contida entre a data de início e de término da turma correspondente.\n- **Unicidade de Data**: Não é permitido agendar duas aulas na mesma data para a mesma turma.",
        "operationId": "AulasController_atualizar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtualizarAulaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "O registro da aula foi atualizado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AulaRespostaDto"
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação de regras de negócio ao atualizar a aula.",
            "content": {
              "application/json": {
                "examples": {
                  "aula_com_chamada_nao_pode_agendar": {
                    "summary": "Tentativa de reverter para AGENDADA com chamada existente",
                    "value": {
                      "statusCode": 400,
                      "codigo": "AULA_COM_CHAMADA_NAO_PODE_AGENDAR",
                      "message": "Esta aula já possui registros de presença lançados e não pode retornar ao status de AGENDADA. Para reverter o status, exclua a lista de chamadas da aula primeiro.",
                      "error": "Bad Request"
                    }
                  },
                  "aula_com_chamada_nao_pode_cancelar": {
                    "summary": "Tentativa de cancelar aula que já possui lista de chamada",
                    "value": {
                      "statusCode": 400,
                      "codigo": "AULA_COM_CHAMADA_NAO_PODE_CANCELAR",
                      "message": "Não é possível CANCELAR uma aula que já possui registros de chamada salvos no sistema. Exclua a lista de chamadas da aula antes de cancelá-la.",
                      "error": "Bad Request"
                    }
                  },
                  "aula_realizada_sem_chamada": {
                    "summary": "Tentativa de marcar como REALIZADA sem chamada lançada",
                    "value": {
                      "statusCode": 400,
                      "codigo": "AULA_REALIZADA_SEM_CHAMADA",
                      "message": "Não é possível marcar uma aula como REALIZADA sem antes registrar o lote de chamadas dos alunos.",
                      "error": "Bad Request"
                    }
                  },
                  "aula_data_anterior_inicio_turma": {
                    "summary": "Data anterior ao início da turma",
                    "value": {
                      "statusCode": 400,
                      "codigo": "AULA_DATA_ANTERIOR_INICIO_TURMA",
                      "message": "A data da aula não pode ser anterior à data de início da turma.",
                      "error": "Bad Request"
                    }
                  },
                  "aula_data_posterior_fim_turma": {
                    "summary": "Data posterior ao encerramento da turma",
                    "value": {
                      "statusCode": 400,
                      "codigo": "AULA_DATA_POSTERIOR_FIM_TURMA",
                      "message": "A data da aula não pode ser posterior à data de encerramento da turma.",
                      "error": "Bad Request"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Aula com o ID especificado não foi encontrada."
          },
          "409": {
            "description": "Conflito de duplicidade de data na turma.",
            "content": {
              "application/json": {
                "examples": {
                  "aula_data_duplicada": {
                    "summary": "Aula já cadastrada para esta mesma data na turma",
                    "value": {
                      "statusCode": 409,
                      "codigo": "AULA_DATA_DUPLICADA",
                      "message": "Já existe uma aula cadastrada nesta mesma data para esta turma.",
                      "error": "Conflict"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Ocorreu um erro inesperado ao atualizar a aula."
          }
        },
        "summary": "Atualizar dados ou o status de uma aula pelo ID",
        "tags": [
          "Aulas"
        ]
      },
      "delete": {
        "operationId": "AulasController_remover",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "A aula foi removida com sucesso (soft delete)."
          },
          "400": {
            "description": "Esta aula não pode ser removida pois possui um lote de chamadas ativo e vinculado a ela."
          },
          "404": {
            "description": "Aula com o ID especificado não foi encontrada."
          },
          "500": {
            "description": "Ocorreu um erro inesperado ao remover a aula."
          }
        },
        "summary": "Remover logicamente uma aula do cronograma pelo ID",
        "tags": [
          "Aulas"
        ]
      }
    },
...
    "/chamadas/aula/{aulaId}": {
      "get": {
        "operationId": "ChamadasController_buscarPorAula",
        "parameters": [
          {
            "name": "aulaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A lista de chamadas da aula foi recuperada com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChamadaRespostaDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "A aula especificada não possui registros de chamada ou não foi encontrada."
          },
          "500": {
            "description": "Ocorreu um erro inesperado ao buscar as presenças da aula."
          }
        },
        "summary": "Buscar a lista de chamadas preenchida de uma aula específica",
        "tags": [
          "Chamadas (presença nas aulas)"
        ]
      },
      "delete": {
        "description": "Exclui todas as presenças/faltas registradas para a aula especificada (soft delete).\nSe a aula estiver com status REALIZADA, o status é revertido automaticamente para AGENDADA.",
        "operationId": "ChamadasController_removerPorAula",
        "parameters": [
          {
            "name": "aulaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "As presenças da aula foram removidas com sucesso e o status da aula foi revertido para AGENDADA."
          },
          "404": {
            "description": "A aula informada não foi encontrada ou não possui registros de chamada para serem removidos."
          },
          "500": {
            "description": "Ocorreu um erro inesperado ao remover a lista de chamadas."
          }
        },
        "summary": "Remover o lote de chamadas de uma aula (limpar presenças)",
        "tags": [
          "Chamadas (presença nas aulas)"
        ]
      }
    },
...
  },
  "info": {
    "title": "Portal ONG Ser Amor API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "PaginacaoMetaDto": {
        "type": "object",
        "properties": {
          "paginaAtual": {
            "type": "number",
            "description": "Página atual",
            "example": 1
          },
          "itensPorPagina": {
            "type": "number",
            "description": "Número de itens por página",
            "example": 10
          },
          "totalItens": {
            "type": "number",
            "description": "Total de itens",
            "example": 100
          },
          "totalPaginas": {
            "type": "number",
            "description": "Total de páginas",
            "example": 10
          }
        },
        "required": [
          "paginaAtual",
          "itensPorPagina",
          "totalItens",
          "totalPaginas"
        ]
      },
      "PaginacaoRespostaDto": {
        "type": "object",
        "properties": {
          "meta": {
            "description": "Metadados da paginação",
            "allOf": [
              {
                "$ref": "#/components/schemas/PaginacaoMetaDto"
              }
            ]
          },
          "dados": {
            "description": "Dados paginados",
            "type": "array",
            "items": {
              "type": "array"
            }
          }
        },
        "required": [
          "meta",
          "dados"
        ]
      },
...
      "CriarAulaDto": {
        "type": "object",
        "properties": {
          "turmaId": {
            "type": "string",
            "example": "1"
          },
          "data": {
            "type": "string",
            "example": "2026-09-26",
            "description": "Data de realização da aula no formato YYYY-MM-DD"
          },
          "tema": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255,
            "example": "Primeira aula",
            "description": "Conteúdo programático, tema ou assunto abordado no encontro"
          }
        },
        "required": [
          "turmaId",
          "data",
          "tema"
        ]
      },
      "AulaRespostaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "turmaId": {
            "type": "string",
            "example": "1"
          },
          "data": {
            "type": "string",
            "example": "2026-09-26"
          },
          "tema": {
            "type": "string",
            "example": "Aula 3"
          },
          "status": {
            "enum": [
              "AGENDADA",
              "REALIZADA",
              "CANCELADA"
            ],
            "type": "string",
            "example": "AGENDADA"
          }
        },
        "required": [
          "id",
          "turmaId",
          "data",
          "tema",
          "status"
        ]
      },
      "AtualizarAulaDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "string",
            "example": "2026-09-26",
            "description": "Data de realização da aula no formato YYYY-MM-DD"
          },
          "tema": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255,
            "example": "Primeira aula",
            "description": "Conteúdo programático, tema ou assunto abordado no encontro"
          },
          "status": {
            "enum": [
              "AGENDADA",
              "REALIZADA",
              "CANCELADA"
            ],
            "type": "string",
            "example": "REALIZADA",
            "description": "Atualiza a situação atual da aula no diário de classe."
          }
        }
      },
...
      "MatriculaResumoRespostaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "beneficiarioId": {
            "type": "string"
          },
          "nomeAluno": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "beneficiarioId",
          "nomeAluno"
        ]
      },
      "ChamadaRespostaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "aulaId": {
            "type": "string"
          },
          "matriculaId": {
            "type": "string"
          },
          "presente": {
            "type": "boolean"
          },
          "faltaJustificada": {
            "type": "boolean"
          },
          "motivoJustificativa": {
            "nullable": true,
            "enum": [
              "SAUDE",
              "OUTROS"
            ],
            "type": "string"
          },
          "observacao": {
            "type": "string",
            "nullable": true
          },
          "matricula": {
            "$ref": "#/components/schemas/MatriculaResumoRespostaDto"
          }
        },
        "required": [
          "id",
          "aulaId",
          "matriculaId",
          "presente",
          "faltaJustificada",
          "motivoJustificativa",
          "observacao"
        ]
      },
...
    }
  }
}